@@ -15,17 +15,17 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.6</version>
+        <version>3.0.1</version>
     </parent>
 
     <properties>
         <jar.finalName>elide-spring-boot</jar.finalName>
         <postgres.version>42.3.7</postgres.version>
-        <liquibase.version>4.8.0</liquibase.version>
-        <elide.version>6.1.4</elide.version>
-        <spring.version>2.5.6</spring.version>
-        <junit.jupiter.version>5.5.2</junit.jupiter.version>
-        <java.version>11</java.version>
+        <liquibase.version>4.18.0</liquibase.version>
+        <elide.version>6.1.11-SNAPSHOT</elide.version>
+        <spring.version>3.0.1</spring.version>
+        <junit.jupiter.version>5.9.1</junit.jupiter.version>
+        <java.version>17</java.version>
     </properties>
 
     <dependencies>
@@ -34,7 +34,7 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.10</version>
+            <version>1.18.24</version>
         </dependency>
 
         <dependency>
@@ -61,6 +61,11 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-jetty</artifactId>
         </dependency>
+            <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+            <version>5.0.0</version>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -70,13 +75,13 @@
         <dependency>
             <groupId>org.apache.activemq</groupId>
             <artifactId>artemis-jms-server</artifactId>
-            <version>2.18.0</version>
+            <version>2.27.1</version>
         </dependency>
 
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-sleuth</artifactId>
-            <version>3.0.4</version>
+            <version>3.1.5</version>
         </dependency>
 
         <dependency>
@@ -88,7 +93,7 @@
         <dependency>
             <groupId>com.h2database</groupId>
             <artifactId>h2</artifactId>
-            <version>2.1.210</version>
+            <version>2.1.214</version>
         </dependency>
 
         <dependency>
@@ -113,9 +118,9 @@
         </dependency>
 
         <dependency>
-            <groupId>com.jayway.restassured</groupId>
+            <groupId>io.rest-assured</groupId>
             <artifactId>rest-assured</artifactId>
-            <version>2.9.0</version>
+            <version>5.3.0</version>
             <scope>test</scope>
         </dependency>
 
@@ -154,7 +159,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.22.2</version>
+                <version>3.0.0-M7</version>
             </plugin>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
